# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Resume OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /resumes:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List all resumes
      responses:
        "200":
          description: List of resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"
    post:
      operationId: createResume
      tags: [resumes]
      summary: Create a new resume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResumeBody"
      responses:
        "201":
          description: Created resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
  /resumes/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a resume by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateResume
      tags: [resumes]
      summary: Update a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResumeBody"
      responses:
        "200":
          description: Updated resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Resume:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        template:
          type: string
        profile:
          type: object
          additionalProperties: true
        experience:
          type: array
          items:
            type: object
            additionalProperties: true
        education:
          type: array
          items:
            type: object
            additionalProperties: true
        skills:
          type: array
          items:
            type: object
            additionalProperties: true
        projects:
          type: array
          items:
            type: object
            additionalProperties: true
        certifications:
          type: array
          items:
            type: object
            additionalProperties: true
        atsScore:
          type: integer
          nullable: true
        targetRole:
          type: string
          nullable: true
        targetIndustry:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - template
        - profile
        - experience
        - education
        - skills
        - projects
        - certifications
        - createdAt
        - updatedAt
    CreateResumeBody:
      type: object
      properties:
        title:
          type: string
        template:
          type: string
      required:
        - title
    UpdateResumeBody:
      type: object
      properties:
        title:
          type: string
        template:
          type: string
        profile:
          type: object
          additionalProperties: true
        experience:
          type: array
          items:
            type: object
            additionalProperties: true
        education:
          type: array
          items:
            type: object
            additionalProperties: true
        skills:
          type: array
          items:
            type: object
            additionalProperties: true
        projects:
          type: array
          items:
            type: object
            additionalProperties: true
        certifications:
          type: array
          items:
            type: object
            additionalProperties: true
        targetRole:
          type: string
        targetIndustry:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error