# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume OS API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListAnthropicConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
});
export const ListAnthropicConversationsResponse = zod.array(
  ListAnthropicConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateAnthropicConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetAnthropicConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAnthropicConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteAnthropicConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListAnthropicMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListAnthropicMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.date(),
});
export const ListAnthropicMessagesResponse = zod.array(
  ListAnthropicMessagesResponseItem,
);

/**
 * @summary Send a message and receive an AI response (SSE stream)
 */
export const SendAnthropicMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendAnthropicMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary List all resumes
 */
export const ListResumesResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  template: zod.string(),
  profile: zod.record(zod.string(), zod.unknown()),
  experience: zod.array(zod.record(zod.string(), zod.unknown())),
  education: zod.array(zod.record(zod.string(), zod.unknown())),
  skills: zod.array(zod.record(zod.string(), zod.unknown())),
  projects: zod.array(zod.record(zod.string(), zod.unknown())),
  certifications: zod.array(zod.record(zod.string(), zod.unknown())),
  atsScore: zod.number().nullish(),
  targetRole: zod.string().nullish(),
  targetIndustry: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListResumesResponse = zod.array(ListResumesResponseItem);

/**
 * @summary Create a new resume
 */
export const CreateResumeBody = zod.object({
  title: zod.string(),
  template: zod.string().optional(),
});

/**
 * @summary Get a resume by ID
 */
export const GetResumeParams = zod.object({
  id: zod.coerce.number(),
});

export const GetResumeResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  template: zod.string(),
  profile: zod.record(zod.string(), zod.unknown()),
  experience: zod.array(zod.record(zod.string(), zod.unknown())),
  education: zod.array(zod.record(zod.string(), zod.unknown())),
  skills: zod.array(zod.record(zod.string(), zod.unknown())),
  projects: zod.array(zod.record(zod.string(), zod.unknown())),
  certifications: zod.array(zod.record(zod.string(), zod.unknown())),
  atsScore: zod.number().nullish(),
  targetRole: zod.string().nullish(),
  targetIndustry: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update a resume
 */
export const UpdateResumeParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateResumeBody = zod.object({
  title: zod.string().optional(),
  template: zod.string().optional(),
  profile: zod.record(zod.string(), zod.unknown()).optional(),
  experience: zod.array(zod.record(zod.string(), zod.unknown())).optional(),
  education: zod.array(zod.record(zod.string(), zod.unknown())).optional(),
  skills: zod.array(zod.record(zod.string(), zod.unknown())).optional(),
  projects: zod.array(zod.record(zod.string(), zod.unknown())).optional(),
  certifications: zod.array(zod.record(zod.string(), zod.unknown())).optional(),
  targetRole: zod.string().optional(),
  targetIndustry: zod.string().optional(),
});

export const UpdateResumeResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  template: zod.string(),
  profile: zod.record(zod.string(), zod.unknown()),
  experience: zod.array(zod.record(zod.string(), zod.unknown())),
  education: zod.array(zod.record(zod.string(), zod.unknown())),
  skills: zod.array(zod.record(zod.string(), zod.unknown())),
  projects: zod.array(zod.record(zod.string(), zod.unknown())),
  certifications: zod.array(zod.record(zod.string(), zod.unknown())),
  atsScore: zod.number().nullish(),
  targetRole: zod.string().nullish(),
  targetIndustry: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a resume
 */
export const DeleteResumeParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary AI-improve a resume section (SSE stream)
 */
export const ImproveSectionBody = zod.object({
  resumeId: zod.number().optional(),
  section: zod
    .string()
    .describe(
      "Section name (profile, experience, education, skills, projects)",
    ),
  content: zod
    .record(zod.string(), zod.unknown())
    .describe("Current section content"),
  targetRole: zod.string().nullish(),
  targetIndustry: zod.string().nullish(),
  yearsOfExperience: zod.number().nullish(),
  companyType: zod
    .string()
    .nullish()
    .describe("startup, corporate, faang, consulting"),
  tone: zod
    .string()
    .nullish()
    .describe("concise, detailed, technical, business"),
  metrics: zod
    .string()
    .nullish()
    .describe("Specific metrics\/achievements to include"),
  previousVersion: zod.record(zod.string(), zod.unknown()).nullish(),
});

/**
 * @summary Get ATS score for a resume
 */
export const GetAtsScoreBody = zod.object({
  resumeId: zod.number().optional(),
  jobDescription: zod.string().nullish(),
  resumeContent: zod.record(zod.string(), zod.unknown()),
});

export const GetAtsScoreResponse = zod.object({
  totalScore: zod.number(),
  breakdown: zod.object({
    keywords: zod.number(),
    structure: zod.number(),
    contentStrength: zod.number(),
    readability: zod.number(),
    measurableAchievements: zod.number(),
  }),
  missingKeywords: zod.array(zod.string()),
  suggestions: zod.array(zod.string()),
  sectionScores: zod.record(zod.string(), zod.number()),
});

/**
 * @summary Analyze resume against job description and suggest improvements (SSE stream)
 */
export const TargetJobBody = zod.object({
  resumeId: zod.number().nullish(),
  resumeContent: zod.record(zod.string(), zod.unknown()),
  jobDescription: zod.string(),
  confirmOptimize: zod.boolean().optional(),
});

/**
 * @summary Global AI chat about the resume (SSE stream)
 */
export const ResumeChatBody = zod.object({
  resumeId: zod.number().nullish(),
  resumeContent: zod.record(zod.string(), zod.unknown()).nullish(),
  message: zod.string(),
  conversationHistory: zod
    .array(
      zod.object({
        role: zod.string(),
        content: zod.string(),
      }),
    )
    .optional(),
});