# lib/api-zod/src/generated/types/updateResumeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume OS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateResumeBodyCertificationsItem } from "./updateResumeBodyCertificationsItem";
import type { UpdateResumeBodyEducationItem } from "./updateResumeBodyEducationItem";
import type { UpdateResumeBodyExperienceItem } from "./updateResumeBodyExperienceItem";
import type { UpdateResumeBodyProfile } from "./updateResumeBodyProfile";
import type { UpdateResumeBodyProjectsItem } from "./updateResumeBodyProjectsItem";
import type { UpdateResumeBodySkillsItem } from "./updateResumeBodySkillsItem";

export interface UpdateResumeBody {
  title?: string;
  template?: string;
  profile?: UpdateResumeBodyProfile;
  experience?: UpdateResumeBodyExperienceItem[];
  education?: UpdateResumeBodyEducationItem[];
  skills?: UpdateResumeBodySkillsItem[];
  projects?: UpdateResumeBodyProjectsItem[];
  certifications?: UpdateResumeBodyCertificationsItem[];
  targetRole?: string;
  targetIndustry?: string;
}